Revert "[Bug] Extract embedded YouTube videos on KhanAcademy pages"
