+++ src/scripts/domParsers/videoUtils.ts
@@ -1,16 +1,12 @@
 import {Utils} from "../utils";
 
-import {VideoExtractor} from "./VideoExtractor";
-import {KhanAcademyVideoExtractor} from "./khanAcademyVideoExtractor";
-
 export module VideoUtils {
 	export const youTubeWatchVideoBaseUrl = "https://www.youtube.com/watch";
 	export const youTubeVideoIdQueryKey = "v";
 
 	export enum SupportedVideoDomains {
 		YouTube,
-		Vimeo,
-		KhanAcademy
+		Vimeo
 	}
 
 	/**
@@ -33,7 +29,73 @@ export module VideoUtils {
 		return;
 	}
 
-	export function matchRegexFromPageContent(pageContent: string, regexes: RegExp[]): string[] {
+	/**
+	 * Return valid iframe src attribute value for the supported YouTube domain
+	 */
+	export function getYouTubeVideoSrcValue(pageUrl: string): string {
+		if (Utils.isNullOrUndefined(pageUrl)) {
+			return;
+		}
+
+		let youTubeVideoId = getYouTubeVideoId(pageUrl);
+		if (Utils.isNullOrUndefined(youTubeVideoId)) {
+			return;
+		}
+		return "https://www.youtube.com/embed/" + youTubeVideoId;
+	}
+
+	/**
+	 * Return valid iframe src attribute value for the supported Vimeo domain
+	 */
+	export function getVimeoVideoSrcValues(pageContent: string): string[] {
+		if (Utils.isNullOrUndefined(pageContent)) {
+			return;
+		}
+
+		let vimeoIds = getVimeoVideoIds(pageContent);
+		if (Utils.isNullOrUndefined(vimeoIds)) {
+			return;
+		}
+
+		let values = [];
+		for (let id of vimeoIds) {
+			values.push("https://player.vimeo.com/video/" + id);
+		}
+
+		return values;
+	}
+
+	/**
+	 * Return id for a video on YouTube.com
+	 */
+	export function getYouTubeVideoId(youTubeUrl: string): string {
+		if (Utils.isNullOrUndefined(youTubeUrl)) {
+			return;
+		}
+
+		let youTubeId;
+		if (Utils.getPathname(youTubeUrl).indexOf("/watch") === 0) {
+			youTubeId = Utils.getQueryValue(youTubeUrl, youTubeVideoIdQueryKey);
+			if (Utils.isNullOrUndefined(youTubeId)) {
+				return;
+			}
+		}
+
+		if (Utils.getPathname(youTubeUrl).indexOf("/embed") === 0) {
+			let youTubeIdMatch = youTubeUrl.match(/youtube\.com\/embed\/(\w+)/);
+			if (Utils.isNullOrUndefined(youTubeIdMatch) || Utils.isNullOrUndefined(youTubeIdMatch[1])) {
+				return;
+			}
+			youTubeId = youTubeIdMatch[1];
+		}
+
+		return youTubeId;
+	}
+
+	/**
+	 * Retrun id for a video on Vimeo.com
+	 */
+	export function getVimeoVideoIds(pageContent: string): string[] {
 		if (Utils.isNullOrUndefined(pageContent)) {
 			return;
 		}
@@ -43,17 +105,16 @@ export module VideoUtils {
 		// 		- ignore casing
 		// 		- ignore possible whitespacing variations between characters
 		// 		- accept the use of either double- or single-quotes around clip_###
+		let regex = /id\s*=\s*("\s*clip_(\d+)\s*"|'\s*clip_(\d+)\s*')/gi;
 		let m;
 		let matches = [];
-		regexes.forEach((regex) => {
-			while (m = regex.exec(pageContent)) {
-				if (m[2]) {
-					matches.push(m[2]);
-				} else {
-					matches.push(m[3]);
-				}
+		while (m = regex.exec(pageContent)) {
+			if (m[2]) {
+				matches.push(m[2]);
+			} else {
+				matches.push(m[3]);
 			}
-		});
+		}
 
 		if (matches.length === 0) {
 			return;
