+++ src/settings/default.json
@@ -34,7 +34,6 @@
 			"youtube\\.com\\/embed\\/(\\w+)",
 			"vimeo\\.com.*?\\/(\\d+)($|\\?|\\#|\\/$)",
 			"vimeo\\.com.*?\\/(\\d+)\\/\\w+:\\w+(\\/|$)",
-			"khanacademy\\.org(.*)\/v\/(.*)",
 			"\\/(\\d{2}|\\d{4})\\/\\d{1,2}\\/",
 			"\\/(\\d{2}|\\d{4})-\\d{1,2}-\\d{1,2}\\/"
 		]
