+++ src/tests/tests.html
@@ -48,9 +48,7 @@
 	<script type="text/javascript" src="contentCapture/bookmarkHelper_tests.js"></script>
 
 	<script type="text/javascript" src="domParsers/domUtils_tests.js"></script>
-	<script type="text/javascript" src="domParsers/khanAcademyVideo_tests.js"></script>
-	<script type="text/javascript" src="domParsers/vimeoVideo_tests.js"></script>
-	<script type="text/javascript" src="domParsers/youtubeVideo_tests.js"></script>	
+	<script type="text/javascript" src="domParsers/videoUtils_tests.js"></script>
 
 	<script type="text/javascript" src="extensions/authenticationHelper_tests.js"></script>
 	<script type="text/javascript" src="extensions/extensionBase_tests.js"></script>
